# server/src/main/java/org/openapitools/api/PersonApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Person;
import org.openapitools.model.PersonsInner;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-11-15T19:48:45.783+01:00[Europe/Berlin]")
@Validated
@Tag(name = "person", description = "the person API")
public interface PersonApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /person
     * Get a list of all Persons
     *
     * @return Success! (status code 200)
     */
    @Operation(
        operationId = "personGet",
        tags = { "person" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success!", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PersonsInner.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/person",
        produces = { "application/json" }
    )
    default ResponseEntity<List<PersonsInner>> personGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "null";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /person/{id}
     * Deleta a Person by ID
     *
     * @param id  (required)
     * @return Successfully deleted a Person (status code 200)
     */
    @Operation(
        operationId = "personIdDelete",
        tags = { "person" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successfully deleted a Person")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/person/{id}"
    )
    default ResponseEntity<Void> personIdDelete(
        @Parameter(name = "id", description = "", required = true) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /person/{id}
     * Get a Person by ID
     *
     * @param id  (required)
     * @return Success Response with Path Parameter (status code 200)
     */
    @Operation(
        operationId = "personIdGet",
        tags = { "person" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success Response with Path Parameter", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PersonsInner.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/person/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<List<PersonsInner>> personIdGet(
        @Parameter(name = "id", description = "", required = true) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "null";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /person/{id}
     * Update a Person by ID
     *
     * @param id  (required)
     * @return Data for Person with {ID} changed! (status code 200)
     */
    @Operation(
        operationId = "personIdPatch",
        tags = { "person" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Data for Person with {ID} changed!", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Person.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/person/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<Person> personIdPatch(
        @Parameter(name = "id", description = "", required = true) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "null";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /person
     * Add a new Person
     *
     * @param person  (optional)
     * @return Successfully added (status code 201)
     */
    @Operation(
        operationId = "personPost",
        tags = { "person" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Successfully added")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/person",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> personPost(
        @Parameter(name = "Person", description = "") @Valid @RequestBody(required = false) Person person
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}